+++ app/components/PageLayout/PageLayout.module.scss
@@ -9,7 +9,7 @@
 
 .header,
 .content {
-    min-width: 1000px;
+    min-width: 1100px;
     width: 100%;
 }
