+++ app/pages/Accounts/AccountDetailsPage/AccountCarousel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AccountCarousel';
