+++ app/styles/config/_variables.scss
@@ -13,3 +13,5 @@ $anim-easing: ease-out;
 
 $form-color-focus: $color-pale-blue;
 $form-transition: $anim-easing $anim-timing;
+
+$account-balance-height: 300px;
